LPS-166403 Don't expose invocation handler
