+++ modules/apps/static/portal-configuration/portal-configuration-metatype-api/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Portal Configuration Metatype API
 Bundle-SymbolicName: com.liferay.portal.configuration.metatype.api
-Bundle-Version: 4.2.2
+Bundle-Version: 5.0.0
 Export-Package:\
 	com.liferay.portal.configuration.metatype.annotations,\
 	com.liferay.portal.configuration.metatype.bnd.util,\
