+++ modules/apps/static/portal-configuration/portal-configuration-metatype-api/src/main/resources/com/liferay/portal/configuration/metatype/util/packageinfo
@@ -1,1 +1,1 @@
-version 1.1.0
+version 2.0.0
